restructure folder and rename files for semantics
