+++ packages/bitcore-wallet-ui/public/assets/bitcoin.png
