+++ packages/bitcore-wallet-ui/src/components/activity/TransactionContainer.tsx
