+++ packages/bitcore-wallet-ui/src/components/activity/TransactionList.tsx
