+++ packages/bitcore-wallet-ui/src/components/footer/BottomNav.tsx
