+++ packages/bitcore-wallet-ui/src/components/notification/Notification.tsx
