+++ packages/bitcore-wallet-ui/src/components/recieve/AddressList.tsx
