+++ packages/bitcore-wallet-ui/src/components/recieve/AddressNavBar.tsx
@@ -2,7 +2,6 @@ import React, { Component } from 'react';
 import { withStyles, createStyles } from '@material-ui/core/styles';
 import Typography from '@material-ui/core/Typography';
 import { Paper } from '@material-ui/core';
-import { WalletHeader } from '../wallet/WalletHeader';
 import InputBase from '@material-ui/core/InputBase';
 import Divider from '@material-ui/core/Divider';
 import IconButton from '@material-ui/core/IconButton';
@@ -10,6 +9,7 @@ import Button from '@material-ui/core/Button';
 import { AppState } from '../../contexts/state';
 import { ActionCreators, store } from '../../index';
 import { connect } from 'react-redux';
+import { WalletHeader } from '../header/WalletHeader';
 
 const styles = createStyles({
   root: {
