+++ packages/bitcore-wallet-ui/src/components/send/QRBox.tsx
