+++ packages/bitcore-wallet-ui/src/components/unlock/UnlockBar.tsx
