+++ packages/bitcore-wallet-ui/src/components/wallets/AppBar.tsx
