+++ packages/bitcore-wallet-ui/src/components/wallets/CreateWallet.tsx
