+++ packages/bitcore-wallet-ui/src/components/wallets/WalletContainer.tsx
