+++ packages/bitcore-wallet-ui/src/components/wallets/WalletList.tsx
